Finished refactoring read radar and added messages upon timer completions and pomodoro cycle counts.
